Added mode and cooldown functionality
The game now checks what mode a player is in when they make a command in order to determine if such a command is permitted. It also checks if a cooldown is currently on the command. Cooldowns can now be placed on commands and remain in effect on a profile even through exiting and reloading the game.