Created buff functionality
Created functions for creating and removing buffs and added buff resuming to the clean up function. Also added a die function, which runs when the player's HP falls to 0, and added descriptions to each of the functions so that the new 'info' command can display a description of an input command.