Fixed various inventory and equipment functionality
Created 'eqp' command, which is a shorthand version of 'equipment'. Defined 'me' command. Fixed 'inspect' so that it adds a new line before barfing out info. Condensed 'inventory' output so that the output bullets are closer together. Swapped 'inv' and 'inventory' roles. Got rid of equipment condition (HP) and their fields in the 'inspect' command.